Extended filtering of hits and simplified keyword selection
Option included for removing specific singleton proteins. Option included for adding specific proteins that are usually singletons and have co-occurring patterns in a genome. Keywords that are used for transitioning in the synteny completion function now use the simplest longest patterns to try the completion. Otherwise unfitting long patterns were tried to be optimized